+++ config/routes.rb
@@ -91,19 +91,12 @@ Rails.application.routes.draw do
 
     resources :groups do
       member do
-        post 'request', to: 'groups#request_to_join'
-        put 'accept_request'
-        get 'invite'
-        patch 'invite', to: 'groups#send_invites', as: 'send_invites'
-        put 'accept_invitation'
-        put 'reject_pending'
         put 'finalize'
-        put 'finalize_membership'
         put 'lock'
         put 'unlock'
-        delete 'leave'
         patch 'make_drawless'
       end
+
       collection do
         resource :suite_assignment, only: %i(new create destroy)
       end
@@ -126,6 +119,21 @@ Rails.application.routes.draw do
     member do
       put 'lock'
       put 'unlock'
+    end
+    resources :memberships do
+      member do
+        post 'request', to: 'memberships#request_to_join'
+        put 'accept_request'
+        put 'accept_invitation'
+        put 'reject_pending'
+        put 'finalize_membership'
+        delete 'leave'
+      end
+
+      collection do
+        get 'invite'
+        post 'invite', to: 'memberships#send_invites', as: 'send_invites'
+      end
     end
     resource :suite_assignment, only: %i(new create destroy)
     resource :room_assignment, only: %i(new create edit) do